% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_correlation.R
\name{ccf_errors}
\alias{ccf_errors}
\title{Use random simulations to estimate undertainty on CCF estimates.}
\usage{
ccf_errors(ts.1, ts.2, tau = NULL, min.pts = 5, local.est = FALSE,
  cov = FALSE, prob = 0.1, nsim = 250, peak.frac = 0.8,
  zero.clip = NULL, one.way = FALSE, method = "iccf",
  use.errors = FALSE, acf.flag = FALSE, chatter = 0)
}
\arguments{
\item{ts.1}{(array or dataframe) data for time series 1 and 2.}

\item{ts.2}{(array or dataframe) data for time series 1 and 2.}

\item{tau}{(array) list of lags at which the CCF is to be evaluated.}

\item{min.pts}{(integer) each DCF bin must contain at least \code{min.pts} correlation coefficients.}

\item{local.est}{(logical) use 'local' (not 'global') means and variances?}

\item{cov}{(logical) if \code{TRUE} then compute covariance, not correlation coefficient.}

\item{prob}{(logical) probability level to use for confidence intervals}

\item{nsim}{(integer) number of FR/RSS simulations to run}

\item{peak.frac}{(float) only include CCF points above \code{peak.frac}*max(ccf) in centroid calculation.}

\item{zero.clip}{(logical) remove pairs of points with exactly zero lag?}

\item{one.way}{(logical) (ICCF only) if TRUE then only interpolar time series 2.}

\item{method}{(string) use \code{"dcf"} or \code{"iccf"} (default).}

\item{use.errors}{(logical) if \code{TRUE} then subtract mean square error from variances.}

\item{acf.flag}{(logical) \code{TRUE} when computing ACF, 
and \code{ts.2 = ts.1}}

\item{chatter}{(integer) set the level of feedback.}
}
\value{
The output is a list containing two data frames: \code{lags} and \code{dists}.
   \item{lags}{a data frame with four columns}
   \item{tau}{time lags}
   \item{dcf}{the DCF values for the input data}
   \item{lower}{the lower limit of the confidence interval}
   \item{upper}{the upper limit of the confidence interval}
   \item{dists}{a data frame with two columns}
   \item{peak.lag}{the peak values from nsim simulations}
   \item{cent.lag}{the centroid values from nsim simulations}
}
\description{
\code{ccf_errors} returns information on the undertainty on CCF estimates.
}
\details{
Computes errors on the CCF estimates using "flux randomisation" 
and "random subset sampling" FR/RSS using the \code{fr_rss} function.
}
\section{Notes}{

For each randomised pair of light curves we compute the CCF. We record the
CCF, the lag at the peak, and the centroid lag (including only points higher
than \code{peak.frac * max(ccf)}). Using \code{nsim} simulations we compute
the \code{(1-p)*100\%} confidence intervals on the CCF values, and the
distribution of the peaks and centroids.
}

\seealso{
\code{\link{cross_correlate}}, \code{\link{fr_rss}}
}
