% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_functions.R
\name{dcf}
\alias{dcf}
\title{Compute the Discrete Correlation Function}
\usage{
dcf(ts.1, ts.2, tau = NULL, min.pts = 5, local.est = FALSE,
  zero.clip = NULL, use.errors = FALSE, cov = FALSE, chatter = 0)
}
\arguments{
\item{ts.1}{(array or dataframe) data for time series 1 and 2.}

\item{ts.2}{(array or dataframe) data for time series 1 and 2.}

\item{tau}{(vector) list of lags at which to compute the CCF.}

\item{min.pts}{(integer) each DCF bin must contain at least \code{min.pts} correlation coefficients.}

\item{local.est}{(logical) use 'local' (not 'global') means and variances?}

\item{zero.clip}{(logical) remove pairs of points with exactly zero lag?}

\item{use.errors}{(logical) if \code{TRUE} then subtract mean square error from variances.}

\item{cov}{(logical) if \code{TRUE} then compute covariance, not correlation coefficient.}

\item{chatter}{(integer) set the level of feedback.}
}
\value{
A data frame containing columns:
 \item{tau}{the centre of the lag bins (vector)}
 \item{ccf}{the correlation coefficent in each lag bin}
 \item{n}{the number of data point pairs included in each lag bin}
}
\description{
\code{dcf} returns the Discrete Correlation Function estimates.
}
\section{Notes}{

Input are two time series (data frames with columns: \code{t, y, dy}
[optional]) Output is the correlation coefficient \code{ccf[i]} in different
lag bins \code{tau[i]}.

In what follows we refer to the \code{t, y} values of time series 1
(\code{ts.1}) as \code{t.1, y.1}, and similarly for time series 2.

We first subtract the mean values from \code{y.1} and \code{y.2}. Then, 
within the \code{i}th lag bin (\code{tau[i] - dtau/2}, \code{tau[i] +
dtau/2}) we collect all pairs (\code{y.1, y.2}) of data for which \code{t.1 -
t.2} falls within the lag bin. Using these pairs of data we compute the sum
of their product and normalise it:

\code{cov[i] = (1/n) * sum_{j,k=1..n} (y.1[j] * y.2[k])}

Here, \code{j, k} are index arrays of length \code{n} that specify the points
of time series 1 and 2 (respectively) which pair-up within lag bin \code{i}. 
This gives a covariance. If \code{cov = TRUE} we keep these values. Otherwise
(default: \code{cov = FALSE}) we normalise by the product of the standard 
deviations of \code{y.1} and \code{y.2}.

The number of output lags \code{tau} is \code{2*lag.bins+1}, and the lags are
centred on zero. So they run from \code{-max.lag} to \code{+max.lag}. Any lag
bins containing fewer than min.pts pairs of points will have \code{ccf = NA}.

If 'errors' are suppled for either or both time series (dx.1, dx.2) 
and \code{use.errors = TRUE} then the variance used 
in the denominator terms of correlation coefficient calculation will be
the 'excess variance', i.e. the total sample variance minus the 
mean square error. If dy.1 and/or dy.2 are a single number, this is 
assumed to be the same error for each data point. If \code{use.errors = FALSE}
(default) then the usual sample variance will be used.
}

\examples{
## Example using NGC 5548 data
res <- dcf(cont, hbeta, tau = seq(-200, 200, by = 5))
plot(res$tau, res$ccf, type = "l", col = "blue", lwd = 3, bty = "n")
grid()

}
\seealso{
\code{\link{cross.correlate}}, \code{\link{iccf}}
}
