% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccf_functions.R
\name{iccf}
\alias{iccf}
\title{Estimate an interpolated cross correlation between two time series}
\usage{
iccf(ts.1, ts.2, tau = NULL, local.est = FALSE, one.way = FALSE,
  zero.clip = NULL, cov = FALSE, chatter = 0)
}
\arguments{
\item{ts.1}{(array or dataframe) data for time series 1 and 2.}

\item{ts.2}{(array or dataframe) data for time series 1 and 2.}

\item{tau}{(vector) list of lags at which to compute the CCF.}

\item{local.est}{(logical) use 'local' (not 'global') means and variances?}

\item{one.way}{(logical) (ICCF only) if TRUE then only interpolar time series 2.}

\item{zero.clip}{(logical) remove pairs of points with exactly zero lag?}

\item{cov}{(logical) if \code{TRUE} then compute covariance, not correlation coefficient.}

\item{chatter}{(integer) set the level of feedback.}
}
\value{
A data frame containing columns:
 \item{tau}{(array) lags (in time units)}
 \item{ccf}{(array) correlations coefficent in each lag bin}
 \item{n}{(array) A one dimensional array containing the number of pairs of points used at each lag.}
}
\description{
\code{iccf} returns the Interpolated Cross-Correlation Function estimates.
}
\section{Notes}{

In what follows we refer to the \code{t, y} values of time series 1 
(\code{ts.1}) as \code{t.1, y.1}, and similarly for time series 2.

Given two time series \code{y.1} and \code{y.2}, sampled at times \code{t.1}
and \code{t.2}, we estimate a cross correlation function (CCF) by 
interpolating (\code{t.2, y.2}). For a lag \code{tau} we estimate \code{y.2} 
at each time \code{t.1+tau} by interpolating between the two nearest points 
of \code{y.2}. We then pair the values \code{y.1} with the corresponding 
lagged values of \code{y.2} and compute the linear correlation coefficient, 
\code{ccf}. The interpolation is handled by the \code{approx} function for
linear interpolation.
}

\examples{
## Example using NGC 5548 data
res <- iccf(cont, hbeta, tau = seq(-100, 100))
plot(res$tau, res$ccf, type = "l", col = "blue", lwd = 3, bty = "n")
grid()

}
\seealso{
\code{\link{cross.correlate}}, \code{\link{dcf}}, \code{\link[stats]{approx}}
}
