% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccf_functions.R
\name{iccf_core}
\alias{iccf_core}
\title{Compute the one-way Interpolated Cross-Correlation Function (ICCF)}
\usage{
iccf_core(t.1, x.1, t.2, x.2, tau, local.est = FALSE, cov = FALSE)
}
\arguments{
\item{t.1, x.1}{time and value for time series 1}

\item{t.2, x.2}{time and value for time series 2}

\item{local.est}{(logical) use 'local' (not 'global') means and variances?}

\item{cov}{(logical) if \code{TRUE} then compute covariance, not correlation coefficient.}
}
\value{
A list with components
 \item{r}{(array) A one dimensional array containing the correlation
 coefficients at each lag.} 
 \item{n}{(array) A one dimensional array containing the number of pairs of
 points used at each lag.}
}
\description{
\code{iccf_core} returns the basic interpolated correlation coefficients.
}
\details{
The main loop for the ICCF. In this part we take time series 1, \code{x.1} at
\code{t.1}, pair them with values from time series 2, \code{x.2} at
\code{t.1-tau[i]} produce by linearly interpolating between the nearest
values of \code{x.2}. At a given \code{tau[i]} we sum the product of the
paired \code{x.1} and \code{x.2} values \code{r[i] = (1/n) * sum(x.1 * x.2) /
(sd.1 * sd.2)} In the simplest case \code{n}, \code{sd.1} and \code{sd.2} are
constant and are the number of pairs at \code{lag=0} and the total
\code{sqrt(var)} of each time series. If \code{local.est = TRUE} then
\code{n}, \code{sd.1} and \code{sd.2} are evaluated 'locally' i.e. they are
vary for each lag \code{tau[i]}. In this case they are the number of good
pairs at lag \code{tau[i]}, and the \code{sqrt(vars)} of just the \code{x.1}
and \code{x.2} data points involved. We assume \code{x.1} and \code{x.2} have
zero sample mean.
}
\section{Notes}{

 We assume that the input data \code{x.1} and \code{x.2} have been
 mean-subtracted.
}

\examples{
## Example using NGC 5548 data
t1 <- cont$t
y1 <- cont$y - mean(cont$y)
t2 <- hbeta$t
y2 <- hbeta$y - mean(hbeta$y)
tau <- seq(-150, 150)
result <- iccf_core(t1, y1, t2, y2, tau = tau)
plot(tau, result$r, type = "l")
 
}
\seealso{
\code{\link{cross_correlate}}, \code{\link{iccf}}
}
