% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_correlation.R
\name{plot_ccf}
\alias{plot_ccf}
\title{Plot CCF output from cross_correlate}
\usage{
plot_ccf(ccf, ...)
}
\arguments{
\item{ccf}{A list containing \code{tau} and \code{ccf} values as produced by \code{cross_correlate}.}

\item{...}{Any other arguments for the \code{plot()} function.}
}
\value{
None.
}
\description{
\code{plot_ccf} generates a summary plot of CCF data.
}
\seealso{
\code{\link{cross_correlate}}
}
